+++ config.inc.test.php
@@ -23,17 +23,17 @@ define('DB_DB', 'foodsharing');
 define('INFLUX_DSN', 'influxdb://influxdb:8086/foodsharing');
 define('ERROR_REPORT', E_ALL);
 define('BASE_URL', $protocol . '://lmr.local/');
-define('DEFAULT_EMAIL', 'noreply@lebensmittelretten.de');
-define('SUPPORT_EMAIL', 'it@lebensmittelretten.de');
-define('DEFAULT_EMAIL_NAME', 'Foodsharing Freiwillige');
+define('DEFAULT_EMAIL', 'noreply@foodsharing.de');
+define('SUPPORT_EMAIL', 'it@foodsharing.network');
+define('DEFAULT_EMAIL_NAME', 'foodsharing-Freiwillige');
 define('VERSION', '0.8.1');
-define('EMAIL_PUBLIC', 'info@lebensmittelretten.de');
-define('EMAIL_PUBLIC_NAME', 'Foodsharing Freiwillige');
-define('DEFAULT_EMAIL_HOST', 'lebensmittelretten.de');
-define('MAPZEN_API_KEY', 'mapzen-RaXru7A');
+define('EMAIL_PUBLIC', 'info@foodsharing.de');
+define('EMAIL_PUBLIC_NAME', 'foodsharing-Freiwillige');
+define('NOREPLY_EMAIL_HOST', 'foodsharing.de');
+define('PLATFORM_MAILBOX_HOST', 'foodsharing.network');
 define('GOOGLE_API_KEY', '');
 
-define('MAILBOX_OWN_DOMAINS', array('lebensmittelretten.de'));
+define('MAILBOX_OWN_DOMAINS', array('foodsharing.network', 'lebensmittelretten.de', 'foodsharing.de'));
 
 define('SMTP_HOST', 'maildev');
 define('SMTP_PORT', 25);